backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0

# AI/OpenAI
openai==1.3.5

# Environment Variables
python-dotenv==1.0.0

# Email Validation
email-validator==2.1.0

# Optional: For PostgreSQL support
# psycopg2-binary==2.9.9